app/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
Jinja2==3.1.2
Werkzeug==2.3.7
